Adding first pass implementation of kseq from cjfields
Some mods from the original version. We now have changed the package to
Bio::DB::HTS::Kseq, there is a record object with methods on there, the
iterator returns these by default (there is an old school hash version
available) and you must use the Kseq package (not to do so makes for
some odd importing).