Implement -filter for features(-type=>'coverage') / bami_coverage()
In bami_coverage(), use an iterator-based loop rather than hts_fetch()
and callbacks, as there are now several items to pass down to a callback.
As hts_plbuf_push() eventually pushes a copy of the bam1_t, the loop can
reuse a single bam1_t and a single Perl reference to it without concern.

Note that the items being filtered are *reads*, rather than pileup
positions or similar.