Ensure hts_idx_destroy() happens before corresponding hts_close()
In the current HTSlib CRAM implementation, hts_idx_t is really just
a small fake struct pointing at the real index inside the cram_fd.
So it is vital to ensure that each hts_idx_t is destroyed *before*
its corresponding htsFile is closed. Arrange for the htsFile SV's
refcount to be incremented while each hts_idx_t exists so that
Perl's garbage collector respects this order of destruction.

Fixes #67.